fixed user model deleted unused files
