+++ backend/src/models/User.ts
@@ -1,11 +1,10 @@
-import { model, Schema } from 'mongoose';
+import mongoose, { model, Schema } from 'mongoose';
 import isEmail from 'validator/lib/isEmail';
 import { isCohortValid, imageOrTempIdRegex } from '../helpers/validate-url';
 import NotFoundError from '../errors/not-found-error';
 import ErrorMessages from '../helpers/error-messages';
-import { reactionSchema } from './Reaction';
 import {
-  IBlock, ICity, IInfo, IProfile, IUser, IUserModel,
+  IBlock, ICity, IEmotion, IInfo, IProfile, IStudent, IUser, IUserModel, IСomment,
 } from '../types/user-model';
 
 const blockSchema = new Schema<IBlock>(
@@ -84,6 +83,30 @@ const profileSchema = new Schema<IProfile>(
   { _id: false },
 );
 
+const studentSchema = new Schema<IStudent>(
+  {
+    name: {
+      type: String,
+    },
+    email: {
+      type: String,
+    },
+    _id: { type: mongoose.SchemaTypes.ObjectId, ref: 'User' },
+  },
+  { _id: false },
+);
+
+export const reactionSchema = new Schema(
+  {
+    from: studentSchema,
+    target: {
+      type: String,
+      enum: ['hobby', 'status', 'job', 'edu', 'quote', null],
+    },
+  },
+  { discriminatorKey: 'type' },
+);
+
 const userSchema = new Schema<IUser, IUserModel>(
   {
     createdAt: {
@@ -120,6 +143,32 @@ const userSchema = new Schema<IUser, IUserModel>(
   {
     timestamps: { currentTime: () => Math.floor(Date.now() / 1000) },
   },
+);
+
+const docArray = userSchema.path<Schema.Types.Array>('reactions');
+
+const textSchema = new Schema({
+  text: {
+    type: String,
+    maxlength: 200,
+  },
+});
+
+const emotionSchema = new Schema({
+  emotion: {
+    type: String,
+    maxlength: 200,
+  },
+});
+
+export const Text = docArray.discriminator<IСomment>(
+  'Text',
+  textSchema,
+);
+
+export const Emotion = docArray.discriminator<IEmotion>(
+  'Emotion',
+  emotionSchema,
 );
 
 userSchema.static('findUserByEmail', function findUserByEmail(email: string) {
